Ensure a must point at end of step is written
Prior to this commit, when the final time point of a step was also a
must point, and AutoTimeStep() happened to choose a time step that
matched the must point within the tolerance window,
FETimeStepController::CheckMustPoints() would not adjust the time point
and hence not flag the converged major iteration as a must point via
m_nmust.  Under plot level PLOT_MUST_POINTS, FEBioModel::Write checks
m_nmust to decide whether or not to write a time point to the plotfile.
Hence the final must point could be omitted from the plotfile, which
beyond simply being annoying will break optimization by making the cost
function unstable (non-smooth) and erroneous.

This commit ensures any must point at the end of a step is written by
counting the number of user-requested must points up front and storing
the count in FEAnalysis.m_mpcount.  FEBioModel::Write() now checks if
the count of must points maintained by
FETimeStepController::CheckMustPoints() has met the expected count.  If
on CB_STEP_SOLVED the FETimeStepController's count is too small,
FEBioModel::Write() writes the last time point to the plotfile under the
expectation that the problem described above was responsible.

This is not an ideal solution and should be considered temporary.  The
corrective action relies on an assumption.  There should instead be a
checked invariant.  Ultimately, it will be necessary to refactor these
classes to explicitly communicate the type of time point involved in
each major iteration.